Adjust payment-dummy-driver to new api (#403)
