+++ core/payment-driver/dummy/src/lib.rs
@@ -1,4 +1,3 @@
-use crate::{utils, PaymentDriver, PaymentDriverError};
 use chrono::{DateTime, Utc};
 use futures3::lock::Mutex;
 use futures3::prelude::*;
@@ -9,11 +8,15 @@ use std::collections::HashMap;
 use std::future::Future;
 use std::pin::Pin;
 use std::sync::Arc;
+use bigdecimal::BigDecimal;
 use ya_client_model::NodeId;
-use ya_core_model::driver::{
-    AccountBalance, AccountMode, Balance, Currency, PaymentAmount, PaymentConfirmation,
-    PaymentDetails, PaymentStatus,
-};
+use ya_core_model::driver::{AccountMode, PaymentConfirmation, PaymentDetails};
+use ya_payment_driver::{utils, PaymentDriver, PaymentDriverError};
+use ya_payment_driver::processor::PaymentDriverProcessor;
+use ya_persistence::executor::DbExecutor;
+use ya_service_api_interfaces::Provider;
+
+mod service;
 
 #[derive(Clone)]
 pub struct DummyDriver {
@@ -54,61 +57,36 @@ impl PaymentDriver for DummyDriver {
     fn get_account_balance<'a>(
         &'a self,
         _address: &str,
-    ) -> Pin<Box<dyn Future<Output = Result<AccountBalance, PaymentDriverError>> + 'static>> {
+    ) -> Pin<Box<dyn Future<Output = Result<BigDecimal, PaymentDriverError>> + 'static>> {
         let amount = "1000000000000000000000000";
-        Box::pin(future::ready(Ok(AccountBalance {
-            base_currency: Balance {
-                currency: Currency::Gnt,
-                amount: utils::str_to_big_dec(&amount).unwrap(),
-            },
-            gas: Some(Balance {
-                currency: Currency::Eth,
-                amount: utils::str_to_big_dec(&amount).unwrap(),
-            }),
-        })))
+        Box::pin(future::ready(Ok(utils::str_to_big_dec(&amount).unwrap())))
     }
 
     fn schedule_payment<'a>(
         &self,
-        invoice_id: &str,
-        amount: PaymentAmount,
+        amount: BigDecimal,
         sender: &str,
         recipient: &str,
         _due_date: DateTime<Utc>,
-    ) -> Pin<Box<dyn Future<Output = Result<(), PaymentDriverError>> + 'a>> {
+    ) -> Pin<Box<dyn Future<Output = Result<String, PaymentDriverError>> + 'a>> {
         let payments = self.payments.clone();
         let details = PaymentDetails {
             recipient: recipient.to_string(),
             sender: sender.to_string(),
-            amount: amount.base_currency_amount,
+            amount: amount,
             date: Some(Utc::now()),
         };
-        let invoice_id = invoice_id.to_string();
+        // FIXME: generate payment_order_id?
+        let payment_order_id = Arc::strong_count(&payments).to_string();
 
         Box::pin(async move {
-            match payments.lock().await.entry(invoice_id.clone()) {
-                Entry::Occupied(_) => Err(PaymentDriverError::PaymentAlreadyScheduled(invoice_id)),
+            match payments.lock().await.entry(payment_order_id.clone()) {
+                // FIXME: need other properties to check if Payment is already Scheduled
+                Entry::Occupied(_) => Err(PaymentDriverError::PaymentAlreadyScheduled(payment_order_id)),
                 Entry::Vacant(entry) => {
                     entry.insert(details);
-                    Ok(())
+                    Ok(payment_order_id)
                 }
-            }
-        })
-    }
-
-    fn get_payment_status(
-        &self,
-        invoice_id: &str,
-    ) -> Pin<Box<dyn Future<Output = Result<PaymentStatus, PaymentDriverError>> + 'static>> {
-        let payments = self.payments.clone();
-        let invoice_id = invoice_id.to_owned();
-
-        Box::pin(async move {
-            match payments.lock().await.get(&invoice_id) {
-                Some(details) => Ok(PaymentStatus::Ok(PaymentConfirmation::from(
-                    serde_json::to_string(details).unwrap().as_bytes(),
-                ))),
-                None => Err(PaymentDriverError::PaymentNotFound(invoice_id)),
             }
         })
     }
@@ -126,11 +104,21 @@ impl PaymentDriver for DummyDriver {
         &'a self,
         _payer: &str,
         _payee: &str,
-    ) -> Pin<Box<dyn Future<Output = Result<Balance, PaymentDriverError>> + 'static>> {
+    ) -> Pin<Box<dyn Future<Output = Result<BigDecimal, PaymentDriverError>> + 'static>> {
         let amount = "1000000000000000000000000";
-        Box::pin(future::ready(Ok(Balance {
-            currency: Currency::Gnt,
-            amount: utils::str_to_big_dec(&amount).unwrap(),
-        })))
+        Box::pin(future::ready(Ok(utils::str_to_big_dec(&amount).unwrap())))
+    }
+}
+
+pub struct PaymentDriverService;
+
+impl PaymentDriverService {
+    pub async fn gsb<Context: Provider<Self, DbExecutor>>(context: &Context) -> anyhow::Result<()> {
+        let db: DbExecutor = context.component();
+        let driver = DummyDriver::new();
+        let processor = PaymentDriverProcessor::new(driver);
+        self::service::bind_service(&db, processor);
+        self::service::subscribe_to_identity_events().await;
+        Ok(())
     }
 }
